detecting hit

+++ contracts/src/lib/moveChecks.cairo
@@ -5,8 +5,8 @@ use array::ArrayTrait;
 #[derive(Drop, Copy)]
 pub struct CharacterPosition {
     pub id: u32,
-    pub x: i16,
-    pub y: i16,
+    pub x: i64,
+    pub y: i64,
     pub max_steps: u32,
     pub current_step: u32,
 }
@@ -16,6 +16,6 @@ pub fn does_collide(character: CharacterPosition) -> bool {
     false
 }
 
-pub fn check_valid_movement(movement_x: i16, movement_y: i16) -> bool {
+pub fn check_valid_movement(movement_x: i64, movement_y: i64) -> bool {
     return true;
 }
